KSCU-7: logical refactor &  added modularity following code review with BED.
